tictactoe_v1.2/requirements.txt
emoji==2.1.0
numpy==1.23.4
python-tictactoe==0.0.7
